#160994766 Implement Authorisation on the Frontend  (#9)
* feat(auth): Add utitlity to check user log-in status

The utitily checks if user data is set in localstorage and verifies if
the token property is valid or not.
- include relvant tests

* feat(auth): implement user authorisation on the client

Make use of ProtectedRoute component to restrict access to pages. Also redirect
unauthorised user to login page.
- include relevant tests

* feat(auth): enable redirecting user to referrer

Save the referrer's location in props and pass the information to the
redirected-to component. This way when user is authorised the login component
can redirect user to initial location.

* feat(auth): set initial authenticated state to false

nitpicks: - resolve merge conflicts
- update App component snapshot